♻️: :: 오타수정
