+++ presentation/src/main/java/component/modifier/clickableSingle.kt
@@ -1,4 +1,4 @@
-package component
+package com.kim.presentation.component.modifier
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.clickable
@@ -7,12 +7,13 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.composed
 
+
 @SuppressLint("ModifierFactoryUnreferencedReceiver")
 fun Modifier.clickable(
     enabled: Boolean = true,
     onClick: () -> Unit
 ) = composed {
-    val multipleEventsCutter = remember { MultipleEtventsCutter.get() }
+    val multipleEventsCutter = remember { MultipleEventsCutter.get() }
     Modifier.clickable(
         enabled = enabled,
         onClick = { multipleEventsCutter.processEvent { onClick() } },
